+++ backend/compact-connect/lambdas/python/common/requirements.txt
@@ -2,13 +2,13 @@
 # This file is autogenerated by pip-compile with Python 3.12
 # by the following command:
 #
-#    pip-compile --no-emit-index-url compact-connect/lambdas/common-python/requirements.in
+#    pip-compile --no-emit-index-url compact-connect/lambdas/python/common/requirements.in
 #
 aws-lambda-powertools==2.43.1
-    # via -r compact-connect/lambdas/common-python/requirements.in
-boto3==1.35.60
-    # via -r compact-connect/lambdas/common-python/requirements.in
-botocore==1.35.60
+    # via -r compact-connect/lambdas/python/common/requirements.in
+boto3==1.35.67
+    # via -r compact-connect/lambdas/python/common/requirements.in
+botocore==1.35.67
     # via
     #   boto3
     #   s3transfer
@@ -18,12 +18,12 @@ jmespath==1.0.1
     #   boto3
     #   botocore
 marshmallow==3.23.1
-    # via -r compact-connect/lambdas/common-python/requirements.in
+    # via -r compact-connect/lambdas/python/common/requirements.in
 packaging==24.2
     # via marshmallow
 python-dateutil==2.9.0.post0
     # via botocore
-s3transfer==0.10.3
+s3transfer==0.10.4
     # via boto3
 six==1.16.0
     # via python-dateutil
