+++ backend/compact-connect/lambdas/python/data-events/.coveragerc
@@ -1,0 +1,10 @@
+[run]
+data_file = ../../../.coverage
+
+omit =
+  */cdk.out/*
+  */smoke-test/*
+  */tests/*
+
+[report]
+skip_empty = true
