+++ backend/compact-connect/lambdas/python/data-events/requirements.in
@@ -1,2 +1,1 @@
-boto3>=1.34, <2
-aws-lambda-powertools>=2.29.1, <3
+# common requirements are managed in the common requirements.in file
