+++ backend/compact-connect/lambdas/python/data-events/requirements.txt
@@ -2,28 +2,5 @@
 # This file is autogenerated by pip-compile with Python 3.12
 # by the following command:
 #
-#    pip-compile --no-emit-index-url compact-connect/lambdas/data-events/requirements.in
+#    pip-compile --no-emit-index-url compact-connect/lambdas/python/data-events/requirements.in
 #
-aws-lambda-powertools==2.43.1
-    # via -r compact-connect/lambdas/data-events/requirements.in
-boto3==1.35.60
-    # via -r compact-connect/lambdas/data-events/requirements.in
-botocore==1.35.60
-    # via
-    #   boto3
-    #   s3transfer
-jmespath==1.0.1
-    # via
-    #   aws-lambda-powertools
-    #   boto3
-    #   botocore
-python-dateutil==2.9.0.post0
-    # via botocore
-s3transfer==0.10.3
-    # via boto3
-six==1.16.0
-    # via python-dateutil
-typing-extensions==4.12.2
-    # via aws-lambda-powertools
-urllib3==2.2.3
-    # via botocore
