+++ backend/compact-connect/lambdas/python/purchases/requirements.txt
@@ -2,10 +2,10 @@
 # This file is autogenerated by pip-compile with Python 3.12
 # by the following command:
 #
-#    pip-compile --no-emit-index-url compact-connect/lambdas/purchases/requirements.in
+#    pip-compile --no-emit-index-url compact-connect/lambdas/python/purchases/requirements.in
 #
 authorizenet==1.1.5
-    # via -r compact-connect/lambdas/purchases/requirements.in
+    # via -r compact-connect/lambdas/python/purchases/requirements.in
 certifi==2024.8.30
     # via requests
 charset-normalizer==3.4.0
