+++ backend/compact-connect/lambdas/python/staff-user-pre-token/requirements.txt
@@ -2,5 +2,5 @@
 # This file is autogenerated by pip-compile with Python 3.12
 # by the following command:
 #
-#    pip-compile --no-emit-index-url compact-connect/lambdas/staff-user-pre-token/requirements.in
+#    pip-compile --no-emit-index-url compact-connect/lambdas/python/staff-user-pre-token/requirements.in
 #
