+++ backend/compact-connect/lambdas/python/staff-users/requirements-dev.txt
@@ -2,11 +2,11 @@
 # This file is autogenerated by pip-compile with Python 3.12
 # by the following command:
 #
-#    pip-compile --no-emit-index-url compact-connect/lambdas/staff-users/requirements-dev.in
+#    pip-compile --no-emit-index-url compact-connect/lambdas/python/staff-users/requirements-dev.in
 #
-boto3==1.35.60
+boto3==1.35.67
     # via moto
-botocore==1.35.60
+botocore==1.35.67
     # via
     #   boto3
     #   moto
@@ -24,7 +24,7 @@ cryptography==43.0.3
 docker==7.1.0
     # via moto
 faker==28.4.1
-    # via -r compact-connect/lambdas/staff-users/requirements-dev.in
+    # via -r compact-connect/lambdas/python/staff-users/requirements-dev.in
 idna==3.10
     # via requests
 jinja2==3.1.4
@@ -39,8 +39,8 @@ markupsafe==3.0.2
     # via
     #   jinja2
     #   werkzeug
-moto[cognitoidp,dynamodb,s3]==5.0.20
-    # via -r compact-connect/lambdas/staff-users/requirements-dev.in
+moto[cognitoidp,dynamodb,s3]==5.0.21
+    # via -r compact-connect/lambdas/python/staff-users/requirements-dev.in
 py-partiql-parser==0.5.6
     # via moto
 pycparser==2.22
@@ -61,7 +61,7 @@ requests==2.32.3
     #   responses
 responses==0.25.3
     # via moto
-s3transfer==0.10.3
+s3transfer==0.10.4
     # via boto3
 six==1.16.0
     # via python-dateutil
