+++ backend/compact-connect/requirements-dev.txt
@@ -18,7 +18,7 @@ charset-normalizer==3.4.0
     # via requests
 click==8.1.7
     # via pip-tools
-coverage[toml]==7.6.5
+coverage[toml]==7.6.7
     # via
     #   -r compact-connect/requirements-dev.in
     #   pytest-cov
@@ -86,7 +86,7 @@ requests==2.32.3
     #   pip-audit
 rich==13.9.4
     # via pip-audit
-ruff==0.7.3
+ruff==0.7.4
     # via -r compact-connect/requirements-dev.in
 six==1.16.0
     # via
