+++ backend/compact-connect/requirements.txt
@@ -8,24 +8,24 @@ attrs==24.2.0
     # via
     #   cattrs
     #   jsii
-aws-cdk-asset-awscli-v1==2.2.211
+aws-cdk-asset-awscli-v1==2.2.212
     # via aws-cdk-lib
 aws-cdk-asset-kubectl-v20==2.1.3
     # via aws-cdk-lib
 aws-cdk-asset-node-proxy-agent-v6==2.1.0
     # via aws-cdk-lib
-aws-cdk-aws-lambda-python-alpha==2.167.0a0
+aws-cdk-aws-lambda-python-alpha==2.169.0a0
     # via -r compact-connect/requirements.in
 aws-cdk-cloud-assembly-schema==38.0.1
     # via aws-cdk-lib
-aws-cdk-lib==2.167.0
+aws-cdk-lib==2.169.0
     # via
     #   -r compact-connect/requirements.in
     #   aws-cdk-aws-lambda-python-alpha
     #   cdk-nag
 cattrs==24.1.2
     # via jsii
-cdk-nag==2.34.0
+cdk-nag==2.34.3
     # via -r compact-connect/requirements.in
 constructs==10.4.2
     # via
