+++ backend/compact-connect/stacks/persistent_stack/data_event_table.py
@@ -53,7 +53,7 @@ class DataEventTable(Table):
             'EventHandler',
             description='License data event handler',
             lambda_dir='data-events',
-            index=os.path.join('handlers.py'),
+            index=os.path.join('handlers', 'data_events.py'),
             handler='handle_data_events',
             environment={'DATA_EVENT_TABLE_NAME': self.table_name, **stack.common_env_vars},
             alarm_topic=alarm_topic,
