+++ backend/multi-account/requirements.txt
@@ -8,7 +8,7 @@ attrs==24.2.0
     # via
     #   cattrs
     #   jsii
-aws-cdk-asset-awscli-v1==2.2.211
+aws-cdk-asset-awscli-v1==2.2.212
     # via aws-cdk-lib
 aws-cdk-asset-kubectl-v20==2.1.3
     # via aws-cdk-lib
@@ -16,7 +16,7 @@ aws-cdk-asset-node-proxy-agent-v6==2.1.0
     # via aws-cdk-lib
 aws-cdk-cloud-assembly-schema==38.0.1
     # via aws-cdk-lib
-aws-cdk-lib==2.167.0
+aws-cdk-lib==2.169.0
     # via -r multi-account/requirements.in
 cattrs==24.1.2
     # via jsii
